ai_chat_vm/requirements.txt

flask
flask-cors
google-generativeai
markdown
chromadb
sentence-transformers
embedding_functions
PdfReader